updated claim screen, now only phase1
