+++ src/components/dashboard/Claim/PhaseOne.js
@@ -1,21 +1,20 @@
 // @flow
 import React from 'react'
-import { Image, View } from 'react-native'
-
+import { View } from 'react-native'
 import numeral from 'numeral'
 
-import Text from '../../common/view/Text'
 import Section from '../../common/layout/Section'
+import Text from '../../common/view/Text'
 import BigGoodDollar from '../../common/view/BigGoodDollar'
 
 import { weiToGd } from '../../../lib/wallet/utils'
-
-import arrowsDown from '../../../assets/arrowsDown.svg'
+import { openLink } from '../../../lib/utils/linking'
 import ButtonBlock from './ButtonBlock'
 
-Image.prefetch(arrowsDown)
+// eslint-disable-next-line require-await
+const openLearnMoreLink = async () => openLink('https://w3.gooddollar.org/learn/ubi')
 
-const ClaimPhaseOne = ({
+const ClaimPhaseZero = ({
   handleClaim,
   faceRecognition,
   styles,
@@ -29,19 +28,72 @@ const ClaimPhaseOne = ({
       <View style={styles.mainTextBorder}>
         <Section.Text>
           <Section.Text color="surface" fontFamily="slab" fontWeight="bold" fontSize={40}>
-            {`Universal\n`}
-          </Section.Text>
-          <Section.Text color="surface" fontFamily="slab" fontWeight="bold" fontSize={40}>
-            {`Basic Income\nFor All`}
+            {entitlement ? `Claim Your\nDaily Share` : `Just a Few More\nHours To Go...`}
           </Section.Text>
         </Section.Text>
       </View>
-      <Section.Row alignItems="center" justifyContent="center" style={[styles.row, styles.subMainText]}>
-        <View style={styles.bottomContainer}>
-          <Text color="#0C263D" fontSize={16} fontWeight={500} fontFamily="Roboto">
-            {`Total money generated today:`}
-          </Text>
-          <Text color="#0C263D" fontSize={30} fontWeight="bold" fontFamily="Roboto">
+      {entitlement > 0 ? (
+        <Section.Row alignItems="center" justifyContent="center" style={[styles.row, styles.subMainText]}>
+          <View style={styles.amountBlock}>
+            <Text color="#0C263D" fontSize={55} fontWeight="bold" fontFamily="Roboto">
+              <BigGoodDollar
+                reverse={true}
+                number={entitlement}
+                formatter={weiToGd}
+                fontFamily="Roboto"
+                bigNumberProps={{
+                  fontFamily: 'Roboto',
+                  fontSize: 50,
+                  color: 'black',
+                  fontWeight: 'bold',
+                  lineHeight: 36,
+                }}
+                bigNumberUnitProps={{
+                  fontFamily: 'Roboto',
+                  fontSize: 45,
+                  color: 'black',
+                  fontWeight: 'medium',
+                  lineHeight: 20,
+                }}
+              />
+            </Text>
+          </View>
+        </Section.Row>
+      ) : null}
+      <View>
+        <Section.Text style={styles.secondTextBlock}>
+          <Section.Text color="surface" fontFamily="Roboto" fontSize={18}>
+            {`GoodDollar is the world’s first experiment\nto create a framework to generate\nUBI on a global scale.\n`}
+            <Text
+              color="surface"
+              style={styles.learnMoreLink}
+              textDecorationLine="underline"
+              fontSize={18}
+              fontWeight="bold"
+              fontFamily="slab"
+              onPress={openLearnMoreLink}
+            >
+              {'Learn More'}
+            </Text>
+          </Section.Text>
+        </Section.Text>
+      </View>
+    </Section.Stack>
+    <ButtonBlock
+      styles={styles}
+      entitlement={entitlement}
+      isCitizen={isCitizen}
+      nextClaim={nextClaim}
+      handleClaim={handleClaim}
+      faceRecognition={faceRecognition}
+      showLabelOnly
+    />
+    <Section.Stack style={styles.moreInfo}>
+      <View style={styles.space} />
+      <Section.Row style={styles.extraInfoStats}>
+        <Text style={styles.extraInfoWrapper} fontSize={16} fontWeight="bold" fontFamily="Roboto">
+          <Section.Text>{`Today\n `}</Section.Text>
+          <Section.Text fontWeight="bold">
             <BigGoodDollar
               style={{ display: 'contents' }}
               reverse={true}
@@ -59,30 +111,16 @@ const ClaimPhaseOne = ({
                 color: 'black',
               }}
             />
-          </Text>
-        </View>
-      </Section.Row>
-      <Section.Row alignItems="center" justifyContent="center" style={[styles.row, styles.subMainText]}>
-        <Image source={arrowsDown} style={styles.arrowsDown} />
-      </Section.Row>
-    </Section.Stack>
-    <ButtonBlock
-      styles={styles}
-      entitlement={entitlement}
-      isCitizen={isCitizen}
-      nextClaim={nextClaim}
-      handleClaim={handleClaim}
-      faceRecognition={faceRecognition}
-    />
-    <Section.Stack style={styles.moreInfo}>
-      <Section.Row style={styles.extraInfoStats}>
-        <Text style={styles.extraInfoWrapper}>
-          <Section.Text fontWeight="bold">{numeral(people).format('0a')} </Section.Text>
-          <Section.Text>good people have claimed today!</Section.Text>
+          </Section.Text>
+          <Section.Text>{` Claimed by `}</Section.Text>
+          <Section.Text fontWeight="bold" color="black">
+            {numeral(people).format('0a')}{' '}
+          </Section.Text>
+          <Section.Text>{` Good People`}</Section.Text>
         </Text>
       </Section.Row>
     </Section.Stack>
   </>
 )
 
-export default ClaimPhaseOne
+export default ClaimPhaseZero
